# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "灵嘚中间件（LingDei-Middleware）是一个基于 Fiber 的 RESTful API 服务，用于提供灵嘚（LingDei）的后端服务。\n注意，有 🦸 标识的接口需要管理员权限才能访问。",
        "title": "灵嘚中间件",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
...
        "/like/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "给视频点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞管理"
                ],
                "summary": "给视频点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/like/check": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "检查是否点赞过",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞管理"
                ],
                "summary": "检查是否点赞过",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LikeStatusResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/like/count": {
            "get": {
                "description": "获取某个视频的点赞数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞管理"
                ],
                "summary": "获取某个视频的点赞数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LikeCountResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.LikeCountResp"
                        }
                    }
                }
            }
        },
        "/like/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞管理"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/like/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取某位用户自己的点赞列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞管理"
                ],
                "summary": "获取某位用户自己的点赞列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LikeListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.LikeListResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Like": {
            "type": "object",
            "required": [
                "user_uuid",
                "uuid",
                "video_uuid"
            ],
            "properties": {
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "$ref": "#/definitions/model.Video"
                },
                "video_uuid": {
                    "type": "string"
                }
            }
        },
        "model.LikeCountResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "model.LikeListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "like_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Like"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.LikeStatusResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "model.OperationResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Video": {
            "type": "object",
            "required": [
                "author_uuid",
                "category_uuid",
                "name",
                "thumbnail_url",
                "timestamp",
                "url",
                "uuid"
            ],
            "properties": {
                "author_uuid": {
                    "type": "string"
                },
                "category_uuid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}